tests: skip tests requiring git or git repo
In some exotic packaging environments like Nixpkgs,
we enforce heavy amounts of sandboxing; we do not
include git by default in that sandbox.

Additionally, in said exotic packaging environments,
we should not always assume we are in a Git repository.
The Nixpkgs GitHub fetcher functions by downloading a
tarball, not the whole Git directory, so the test fails.